+++ test/public-docs.t
@@ -6,6 +6,7 @@ use Test::More ;
 my $readme        = _read_text('README.md') ;
 my $documentation = _read_text('DOCUMENTATION.md') ;
 my $dockerhub     = _read_text('DOCKERHUB.md') ;
+my $changelog     = _read_text('CHANGELOG.md') ;
 
 for my $document (
   [ 'README.md',        $readme ],
@@ -50,6 +51,20 @@ unlike $dockerhub, qr/based on the latest development Perl target/,
   'Docker Hub does not describe the Codex base as an unstable latest target' ;
 unlike $dockerhub, qr/based on the Perl 5\.43 development target/,
   'Docker Hub does not identify only the Codex Perl development series' ;
+
+my $latest_perl_version  = _latest_perl_version() ;
+my $latest_release_date  = _latest_release_date($changelog) ;
+my $documented_codex_row = qr/\| `codex` \| \Q$latest_perl_version\E \|/ ;
+
+for my $document ( [ 'README.md', $readme ], [ 'DOCKERHUB.md', $dockerhub ], )
+{
+  my ( $name, $content ) = @{$document} ;
+  like $content, $documented_codex_row,
+    "$name documents Codex on the latest Perl target" ;
+  cmp_ok _module_capture_date($content), 'ge', $latest_release_date,
+    "$name module inventory was captured for the current release" ;
+}
+
 like _read_text('.public-files'), qr/^DOCKERHUB\.md$/m,
   'public snapshot includes the Docker Hub overview source' ;
 is _section( $dockerhub, 'PERL_TARGETS' ),
@@ -70,6 +85,31 @@ sub _section {
     = $content
     =~ /<!-- \Q${name}\E_START -->\n(.*?)<!-- \Q${name}\E_END -->/s ;
   return defined $section ? $section : q{} ;
+}
+
+sub _latest_perl_version {
+  my $content = _read_text('perl-versions.conf') ;
+  my @versions ;
+  for my $line ( split /\n/, $content )
+  {
+    next if $line =~ /\A\s*(?:#|\z)/ ;
+    my ($version) = split /\|/, $line ;
+    push @versions, $version ;
+  }
+
+  return $versions[-1] ;
+}
+
+sub _latest_release_date {
+  my ($content) = @_ ;
+  my ($date)    = $content =~ /^## \[\d+\.\d+\.\d+\] - (\d{4}-\d{2}-\d{2})/m ;
+  return $date ;
+}
+
+sub _module_capture_date {
+  my ($content) = @_ ;
+  my ($date)    = $content =~ /Versions captured on (\d{4}-\d{2}-\d{2}) / ;
+  return $date ;
 }
 
 sub _read_text {